telescope: display retaddrs
Extends `telescope` command so it displays not only register names, but
also return addresses names (ret0, ret1, ret2, ...).

The names are also set as GDB variables so they can be referenced via
`$retX` similarly as registers are by default.

Additionally, I moved the `retaddr` command from
pwndbg/commands/stack.py to pwndbg/commands/retaddr.py.
