+++ pwndbg/__init__.py
@@ -41,12 +41,12 @@ import pwndbg.commands.probeleak
 import pwndbg.commands.procinfo
 import pwndbg.commands.radare2
 import pwndbg.commands.reload
+import pwndbg.commands.retaddr
 import pwndbg.commands.rop
 import pwndbg.commands.ropper
 import pwndbg.commands.search
 import pwndbg.commands.segments
 import pwndbg.commands.shell
-import pwndbg.commands.stack
 import pwndbg.commands.start
 import pwndbg.commands.telescope
 import pwndbg.commands.theme
